House Price prediction
In this project, I develop a Python program that predicts the price of houses in Boston using a machine learning algorithm called Linear Regression. Linear regression is a linear approach to modeling the relationship between a scalar response (or dependent variable) and one or more explanatory variables (or independent variables).